backend/requirements.txt
Flask
flask-cors
Flask-JWT-Extended
Flask-SQLAlchemy
python-dotenv
Werkzeug

